requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Mail==0.9.1
Flask-SocketIO==5.3.2
python-socketio==5.11.0
eventlet==0.36.1
face-recognition==1.3.0
opencv-python==4.10.0.82
numpy==1.26.4
Pillow==10.0.0
python-dotenv==1.0.0
itsdangerous==2.1.2
apscheduler==3.10.1
Flask-WTF==1.1.1
Werkzeug==2.2.3
